feat: check storage type availability for events persistence
