+++ packages/analytics-js-plugins/src/beaconQueue/index.ts
@@ -31,6 +31,7 @@ import {
   LOCAL_STORAGE,
   validateEventPayloadSize,
 } from '../shared-chunks/common';
+import { getStorageTypeForEventsPersistence } from '../utilities/eventsDelivery';
 
 const pluginName: PluginName = 'BeaconQueue';
 
@@ -126,7 +127,7 @@ const BeaconQueue = (): ExtensionPlugin => ({
         } as QueueOpts,
         queueProcessCallback,
         storeManager,
-        LOCAL_STORAGE,
+        getStorageTypeForEventsPersistence(logger),
         logger,
         (itemData: BeaconQueueItemData[]): number => {
           const currentTime = getCurrentTimeFormatted();
