+++ packages/analytics-js/src/components/capabilitiesManager/CapabilitiesManager.ts
@@ -14,6 +14,7 @@ import { isValidURL } from '@rudderstack/analytics-js-common/utilities/url';
 import { isDefinedAndNotNull } from '@rudderstack/analytics-js-common/utilities/checks';
 import type { IHttpClient } from '@rudderstack/analytics-js-common/types/HttpClient';
 import { isIE11 } from '@rudderstack/analytics-js-common/utilities/detect';
+import { isStorageAvailable } from '@rudderstack/analytics-js-common/utilities/storage';
 import {
   INVALID_POLYFILL_URL_WARNING,
   POLYFILL_SCRIPT_LOAD_ERROR,
@@ -30,7 +31,6 @@ import {
   hasCrypto,
   hasUAClientHints,
   isLegacyJSEngine,
-  isStorageAvailable,
 } from './detection';
 import { detectAdBlockers } from './detection/adBlockers';
 import { debounce } from '../utilities/globals';
