+++ packages/analytics-js/src/components/capabilitiesManager/detection/index.ts
@@ -3,4 +3,3 @@ export { isBrowser, isNode, hasCrypto, hasUAClientHints, hasBeacon } from './bro
 export { getUserAgentClientHint } from './clientHint';
 export { isDatasetAvailable, legacyJSEngineRequiredPolyfills, isLegacyJSEngine } from './dom';
 export { getScreenDetails } from './screen';
-export { isStorageAvailable, isStorageQuotaExceeded } from './storage';
