+++ packages/analytics-js/src/services/StoreManager/storages/CookieStorage.ts
@@ -5,7 +5,7 @@ import type { ILogger } from '@rudderstack/analytics-js-common/types/Logger';
 import { COOKIE_STORAGE } from '@rudderstack/analytics-js-common/constants/storages';
 import { mergeDeepRight } from '@rudderstack/analytics-js-common/utilities/object';
 import { cookie } from '@rudderstack/analytics-js-common/component-cookie';
-import { isStorageAvailable } from '../../../components/capabilitiesManager/detection';
+import { isStorageAvailable } from '@rudderstack/analytics-js-common/utilities/storage';
 import { getDefaultCookieOptions } from './defaultOptions';
 import { defaultLogger } from '../../Logger';
