+++ packages/analytics-js/src/services/StoreManager/storages/sessionStorage.ts
@@ -7,7 +7,7 @@ import { mergeDeepRight } from '@rudderstack/analytics-js-common/utilities/objec
 import { isUndefined } from '@rudderstack/analytics-js-common/utilities/checks';
 import { SESSION_STORAGE } from '@rudderstack/analytics-js-common/constants/storages';
 import type { Nullable } from '@rudderstack/analytics-js-common/types/Nullable';
-import { isStorageAvailable } from '../../../components/capabilitiesManager/detection';
+import { isStorageAvailable } from '@rudderstack/analytics-js-common/utilities/storage';
 import { defaultLogger } from '../../Logger';
 import { getDefaultSessionStorageOptions } from './defaultOptions';
